🔨 refacto: extract client from intents
