+++ src/intents/helpers.js
@@ -15,3 +15,15 @@ export const errorSerializer = (() => {
     deserialize: data => mapErrorProperties(data, new Error(data.message))
   }
 })()
+
+const first = arr => arr && arr[0]
+// In a far future, the user will have to pick the desired service from a list.
+// For now it's our job, an easy job as we arbitrary pick the first service of
+// the list.
+export function pickService(intent, filterServices) {
+  const services = intent.attributes.services
+  const filteredServices = filterServices
+    ? (services || []).filter(filterServices)
+    : services
+  return first(filteredServices)
+}
